+++ CHANGES.rst
@@ -15,6 +15,7 @@ Improvements:
 Bugfix:
  - Send Access Token as a header instead of a url parameter to upload content (#311)
  - Add API CallSoundsManager.startRingingSilently() to fix issue when incoming call sound is disable (vector-im/riot-android#2417)
+ - Use same TxId when resending an event. The eventId is used as a TxId. (vector-im/riot-android#1997)
 
 API Change:
  - Parameter historyVisibility removed from MxSession.createRoom(). It had no effect.